+++ Fraym/Menu/Entity/MenuItemTranslation.php
@@ -63,6 +63,13 @@ class MenuItemTranslation extends \Fraym\Entity\BaseEntity
     protected $externalUrl = false;
 
     /**
+     * @var string $anchor
+     *
+     * @ORM\Column(name="anchor", type="boolean", nullable=false)
+     */
+    protected $anchor = false;
+
+    /**
      * @var string $shortDesc
      *
      * @ORM\Column(name="short_desc", type="string", length=1000, nullable=true)
@@ -136,7 +143,7 @@ class MenuItemTranslation extends \Fraym\Entity\BaseEntity
             $parts = explode('/', $url);
 
             foreach ($parts as &$part) {
-                $part = $route->createSlug($part);
+                $part = $route->createSlug($part, '-', $this->anchor);
             }
 
             $this->url = '/' . trim(implode('/', $parts), '/');
@@ -153,5 +160,6 @@ class MenuItemTranslation extends \Fraym\Entity\BaseEntity
         $locale = $this->getServiceLocator()->get('Fraym\Locale\Locale');
         $this->locale = $locale->getDefaultLocale();
         $this->externalUrl = false;
+        $this->anchor = false;
     }
 }