+++ Public/install.php
@@ -9,6 +9,10 @@ chdir(realpath(dirname(__FILE__). DIRECTORY_SEPARATOR.'..'.DIRECTORY_SEPARATOR))
 
 require 'Fraym/ClassLoader/ClassLoader.php';
 
+define('CACHE_DI_PATH', 'Cache/DI');
+define('CACHE_DOCTRINE_PROXY_PATH', 'Cache/DoctrineProxies');
+define('CACHE_DOCTRINE_MODULE_FILE', 'Cache/doctrine_module_dir.cache');
+
 $classLoader = new Fraym\ClassLoader\ClassLoader();
 $classLoader->register();
 
@@ -16,22 +20,25 @@ error_reporting(-1);
 ini_set("display_errors", 1);
 
 $builder = new \DI\ContainerBuilder();
+$builder->useAnnotations(true);
 $cache = new Doctrine\Common\Cache\ArrayCache();
 $builder->setDefinitionCache($cache);
+$builder->addDefinitions([
+    'db.options' => array(
+        'driver' => '',
+        'user' =>     '',
+        'password' => '',
+        'host' =>     '',
+        'dbname' =>   '',
+        'charset' => ''
+    )
+]);
 $diContainer = $builder->build();
 
 define('GLOBAL_CACHING_ENABLED', false);
 define('APC_ENABLED', false);
 
 $core = $diContainer->get('Fraym\Core');
-
-$diContainer->set('db.options', array('driver' => '',
-                                      'user' =>     '',
-                                      'password' => '',
-                                      'host' =>     '',
-                                      'dbname' =>   '',
-                                      'charset' => '',
-));
 
 $core->init(\Fraym\Core::ROUTE_CUSTOM);