+++ CHANGELOG.md
@@ -1,13 +1,16 @@
-Fraym 0.9.2, 2014-xx-xx
+Fraym 0.9.2, 2016xx-xx
 ----------------------
 - Added image link
-- *TODO* Added dynamic template configurations
+- Added dynamic template configurations
 - *TODO* Fix flexslider configs
 - Set production mode as default on installation
 - *TODO* Imagine version update
 - Image placeholder fix
 - Change editmode styles
 - Remove eval
+- File manager fixes
+- Entity manager fixes
+- Update vendors
 
 
 Fraym 0.9.1, 2014-04-29